appconfiguration: moving the 2020-06-01 resource manager sdk into a versioned folder
