+++ internal/services/appconfiguration/app_configuration_resource.go
@@ -6,13 +6,14 @@ import (
 	"strings"
 	"time"
 
+	configurationstores2 "github.com/hashicorp/terraform-provider-azurerm/internal/services/appconfiguration/sdk/2020-06-01/configurationstores"
+
 	"github.com/Azure/azure-sdk-for-go/services/appconfiguration/mgmt/2020-06-01/appconfiguration"
 	"github.com/hashicorp/go-azure-helpers/response"
 	"github.com/hashicorp/terraform-provider-azurerm/helpers/azure"
 	"github.com/hashicorp/terraform-provider-azurerm/helpers/tf"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/clients"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/location"
-	"github.com/hashicorp/terraform-provider-azurerm/internal/services/appconfiguration/sdk/configurationstores"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/services/appconfiguration/validate"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/tags"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/tf/pluginsdk"
@@ -35,7 +36,7 @@ func resourceAppConfiguration() *pluginsdk.Resource {
 		},
 
 		Importer: pluginsdk.ImporterValidatingResourceId(func(id string) error {
-			_, err := configurationstores.ConfigurationStoreID(id)
+			_, err := configurationstores2.ConfigurationStoreID(id)
 			return err
 		}),
 
@@ -204,7 +205,7 @@ func resourceAppConfigurationCreate(d *pluginsdk.ResourceData, meta interface{})
 
 	name := d.Get("name").(string)
 	resourceGroup := d.Get("resource_group_name").(string)
-	resourceId := configurationstores.NewConfigurationStoreID(subscriptionId, resourceGroup, name)
+	resourceId := configurationstores2.NewConfigurationStoreID(subscriptionId, resourceGroup, name)
 	existing, err := client.Get(ctx, resourceId)
 	if err != nil {
 		if !response.WasNotFound(existing.HttpResponse) {
@@ -215,9 +216,9 @@ func resourceAppConfigurationCreate(d *pluginsdk.ResourceData, meta interface{})
 		return tf.ImportAsExistsError("azurerm_app_configuration", resourceId.ID())
 	}
 
-	parameters := configurationstores.ConfigurationStore{
+	parameters := configurationstores2.ConfigurationStore{
 		Location: azure.NormalizeLocation(d.Get("location").(string)),
-		Sku: configurationstores.Sku{
+		Sku: configurationstores2.Sku{
 			Name: d.Get("sku").(string),
 		},
 		Tags: expandTags(d.Get("tags").(map[string]interface{})),
@@ -239,13 +240,13 @@ func resourceAppConfigurationUpdate(d *pluginsdk.ResourceData, meta interface{})
 	defer cancel()
 
 	log.Printf("[INFO] preparing arguments for Azure ARM App Configuration update.")
-	id, err := configurationstores.ConfigurationStoreID(d.Id())
+	id, err := configurationstores2.ConfigurationStoreID(d.Id())
 	if err != nil {
 		return err
 	}
 
-	parameters := configurationstores.ConfigurationStoreUpdateParameters{
-		Sku: &configurationstores.Sku{
+	parameters := configurationstores2.ConfigurationStoreUpdateParameters{
+		Sku: &configurationstores2.Sku{
 			Name: d.Get("sku").(string),
 		},
 		Tags: expandTags(d.Get("tags").(map[string]interface{})),
@@ -267,7 +268,7 @@ func resourceAppConfigurationRead(d *pluginsdk.ResourceData, meta interface{}) e
 	ctx, cancel := timeouts.ForRead(meta.(*clients.Client).StopContext, d)
 	defer cancel()
 
-	id, err := configurationstores.ConfigurationStoreID(d.Id())
+	id, err := configurationstores2.ConfigurationStoreID(d.Id())
 	if err != nil {
 		return err
 	}
@@ -319,7 +320,7 @@ func resourceAppConfigurationDelete(d *pluginsdk.ResourceData, meta interface{})
 	ctx, cancel := timeouts.ForDelete(meta.(*clients.Client).StopContext, d)
 	defer cancel()
 
-	id, err := configurationstores.ConfigurationStoreID(d.Id())
+	id, err := configurationstores2.ConfigurationStoreID(d.Id())
 	if err != nil {
 		return err
 	}
@@ -338,7 +339,7 @@ type flattenedAccessKeys struct {
 	secondaryWriteKey []interface{}
 }
 
-func flattenAppConfigurationAccessKeys(values []configurationstores.AccessKey) flattenedAccessKeys {
+func flattenAppConfigurationAccessKeys(values []configurationstores2.AccessKey) flattenedAccessKeys {
 	result := flattenedAccessKeys{
 		primaryReadKey:    make([]interface{}, 0),
 		primaryWriteKey:   make([]interface{}, 0),
@@ -375,7 +376,7 @@ func flattenAppConfigurationAccessKeys(values []configurationstores.AccessKey) f
 	return result
 }
 
-func flattenAppConfigurationAccessKey(input configurationstores.AccessKey) []interface{} {
+func flattenAppConfigurationAccessKey(input configurationstores2.AccessKey) []interface{} {
 	connectionString := ""
 
 	if input.ConnectionString != nil {
@@ -401,23 +402,23 @@ func flattenAppConfigurationAccessKey(input configurationstores.AccessKey) []int
 	}
 }
 
-func expandAppConfigurationIdentity(identities []interface{}) *configurationstores.ResourceIdentity {
-	var out = func(in configurationstores.IdentityType) *configurationstores.ResourceIdentity {
-		return &configurationstores.ResourceIdentity{
+func expandAppConfigurationIdentity(identities []interface{}) *configurationstores2.ResourceIdentity {
+	var out = func(in configurationstores2.IdentityType) *configurationstores2.ResourceIdentity {
+		return &configurationstores2.ResourceIdentity{
 			Type: &in,
 		}
 	}
 
 	if len(identities) == 0 {
-		return out(configurationstores.IdentityTypeNone)
+		return out(configurationstores2.IdentityTypeNone)
 	}
 	identity := identities[0].(map[string]interface{})
-	identityType := configurationstores.IdentityType(identity["type"].(string))
+	identityType := configurationstores2.IdentityType(identity["type"].(string))
 	return out(identityType)
 }
 
-func flattenAppConfigurationIdentity(identity *configurationstores.ResourceIdentity) []interface{} {
-	if identity == nil || identity.Type == nil || *identity.Type == configurationstores.IdentityTypeNone {
+func flattenAppConfigurationIdentity(identity *configurationstores2.ResourceIdentity) []interface{} {
+	if identity == nil || identity.Type == nil || *identity.Type == configurationstores2.IdentityTypeNone {
 		return []interface{}{}
 	}
