+++ internal/services/appconfiguration/sdk/2020-06-01/configurationstores/id_configurationstore_test.go
