+++ internal/services/appconfiguration/sdk/2020-06-01/configurationstores/method_delete_autorest.go
