+++ internal/services/appconfiguration/sdk/2020-06-01/configurationstores/model_sku.go
